requirements.txt
Django>=4.2,<6.0
djangorestframework>=3.15
django-filter>=24.2

# Topic 9 (Caching)
django-redis>=5.4
redis>=5.0

# Topic 10 (Async tasks / scheduling)
celery>=5.4

# QR code attendance (Topic requirement)
qrcode>=8.0
pillow>=10.0

# Deployment (Render)
gunicorn>=22.0
whitenoise>=6.7
dj-database-url>=2.2
psycopg2-binary>=2.9
